+++ src/main/java/lk/edu/esoft/alsskillminercloud/repository/BadgeRepository.java
@@ -3,5 +3,5 @@ package lk.edu.esoft.alsskillminercloud.repository;
 import lk.edu.esoft.alsskillminercloud.entity.Badge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BadgeRepository extends JpaRepository<Badge, Integer> {
+public interface BadgeRepository extends JpaRepository<Badge, Long> {
 }
